% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSCconsultations-package.R
\docType{package}
\name{DHSCconsultations-package}
\alias{DHSCconsultations}
\alias{DHSCconsultations-package}
\title{DHSCconsultations: Collection of tools for analysing consultation responses}
\description{
Package contains a collection of functions that can be used to perform LDA analysis of consultation responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DataS-DHSC/DHSCconsultations}
  \item Report bugs at \url{https://github.com/DataS-DHSC/DHSCconsultations/issues}
}

}
\author{
\strong{Maintainer}:  Data Science Team \email{datascience@dhsc.gov.uk}

}
\keyword{internal}
