% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_glossary.R
\name{get_tlap_glossary}
\alias{get_tlap_glossary}
\title{Return glossary terms from www.thinklocalactpersonal.org.uk}
\usage{
get_tlap_glossary()
}
\value{
list of multi-word glossary terms
}
\description{
Return glossary terms from www.thinklocalactpersonal.org.uk
}
