% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_summary.R
\name{plot_ngram_summary}
\alias{plot_ngram_summary}
\title{Helper function to plot frequency graphs of n-gram summary}
\usage{
plot_ngram_summary(ngram_list, plot_rows = 30)
}
\arguments{
\item{ngram_list}{named list of lists where names are n-gram types and
values are named lists where names are questions and values n-gram counts
(generated by \code{\link[=summarise_ngrams]{summarise_ngrams()}}).}

\item{plot_rows}{integer number of terms to display on plots.}
}
\value{
passed \code{ngram_list}
}
\description{
Helper function to plot frequency graphs of n-gram summary
}
