% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_summary.R
\name{plot_word_summary}
\alias{plot_word_summary}
\title{Helper function to plot frequency graphs of word summary}
\usage{
plot_word_summary(count_list, plot_rows = 30)
}
\arguments{
\item{count_list}{named list of data frame where names are questions and
values are word counts (generated by \code{\link[=summarise_words]{summarise_words()}}).}

\item{plot_rows}{integer number of terms to display on plots.}
}
\value{
passed \code{count_list}
}
\description{
Helper function to plot frequency graphs of word summary
}
