% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_word_summary.R
\name{raw_word_summary}
\alias{raw_word_summary}
\title{Title}
\usage{
raw_word_summary(
  data,
  text_cleaner = NULL,
  stop_words = NULL,
  glossary_words = NULL,
  col_id = NULL,
  cols_free_text = NULL
)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
