% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{replace_glossary}
\alias{replace_glossary}
\title{Replace multi-word glossary terms with single word equivalents}
\usage{
replace_glossary(x, glossary)
}
\arguments{
\item{x}{character vector of text to be transformed.}

\item{glossary}{character vector of multi-word character strings to be
replaced.}
}
\value{
character vector with glossary words replaced with their single word
equivalents.
}
\description{
This function allows multi-word terms to be preserved during LDA.
For example, National Health Service becomes "nationalhealthservice" rather
than "national", "health", and "service"
}
