% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_summary.R
\name{summarise_ngrams}
\alias{summarise_ngrams}
\title{Title}
\usage{
summarise_ngrams(
  data,
  text_cleaner,
  n = list(`bi-gram` = 2, `tri-gram` = 3),
  stop_words = NULL,
  glossary_words = NULL,
  col_id = NULL,
  cols_free_text = NULL
)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
