% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steming_summary.R
\name{summarise_stemming}
\alias{summarise_stemming}
\title{Title}
\usage{
summarise_stemming(
  data,
  text_cleaner,
  stem_word_exceptions,
  stop_words = NULL,
  glossary_words = NULL,
  col_id = NULL,
  cols_free_text = NULL
)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
