% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaner_summary.R
\name{summarise_text_cleaner}
\alias{summarise_text_cleaner}
\title{Title}
\usage{
summarise_text_cleaner(
  data,
  text_cleaner,
  col_id = NULL,
  cols_free_text = NULL
)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
