% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaner_summary.R
\name{text_cleaner_summary}
\alias{text_cleaner_summary}
\title{Title}
\usage{
text_cleaner_summary(data, text_cleaner, col_id = NULL, cols_free_text = NULL)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
