% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_model.R
\name{view_lda_convergence}
\alias{view_lda_convergence}
\title{Show plots of LDA model convergence}
\usage{
view_lda_convergence(lda_list)
}
\arguments{
\item{lda_list}{named list of LDA objects generated by the
\code{\link[=run_lda]{run_lda()}} function.}
}
\value{
invisibly returns passed LDA list object
}
\description{
Show plots of LDA model convergence
}
