% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_fit.R
\name{view_lda_fit_loglikelihood}
\alias{view_lda_fit_loglikelihood}
\title{Show plots of LDA fit log-likelihood}
\usage{
view_lda_fit_loglikelihood(lda_fit_list)
}
\arguments{
\item{lda_fit_list}{named list of lists of LDA objects generated by the
\code{\link[=fit_lda]{fit_lda()}} function.}
}
\value{
invisibly returns passed LDA list object
}
\description{
Show plots of LDA fit log-likelihood
}
