% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_summary.R
\name{view_ngram_summary}
\alias{view_ngram_summary}
\title{Helper function to view n-gram summary}
\usage{
view_ngram_summary(ngram_list)
}
\arguments{
\item{ngram_list}{named list of lists where names are n-gram types and
values are named lists where names are questions and values n-gram counts
(generated by \code{\link[=summarise_ngrams]{summarise_ngrams()}}).}
}
\value{
passed \code{ngram_list}
}
\description{
Helper function to view n-gram summary
}
