% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_summary.R
\name{view_word_summary}
\alias{view_word_summary}
\title{Helper function to view word summary}
\usage{
view_word_summary(count_list)
}
\arguments{
\item{count_list}{named list of data frame where names are questions and
values are word counts (generated by \code{\link[=summarise_words]{summarise_words()}}).}
}
\value{
passed \code{count_list}
}
\description{
Helper function to view word summary
}
