% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_summary.R
\name{write_ngram_summary}
\alias{write_ngram_summary}
\title{Title}
\usage{
write_ngram_summary(
  ngram_list,
  folder,
  plot_rows = 30,
  prefix = basename(folder)
)
}
\arguments{
\item{prefix}{}
}
\description{
Title
}
