% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_word_summary.R
\name{write_raw_word_summary}
\alias{write_raw_word_summary}
\title{Helper function to save raw word summary output to folder}
\usage{
write_raw_word_summary(
  count_list,
  folder,
  plot_rows = 30,
  prefix = basename(folder)
)
}
\arguments{
\item{count_list}{named list of data frame where names are questions and
values are word counts (generated by \code{\link[=summarise_raw_words]{summarise_raw_words()}}).}

\item{folder}{path to the folder in which to save (will create if not
present).}

\item{plot_rows}{integer number of terms to display on plots.}

\item{prefix}{prefix to use when naming files (defaults to
\code{basename(folder)}).}
}
\value{
passed \code{count_list}
}
\description{
Helper function to save raw word summary output to folder
}
