% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_model.R
\name{run_lda_dtm}
\alias{run_lda_dtm}
\title{Calculate LDA topic model}
\usage{
run_lda_dtm(dtm, k, seed, burn_in = 1000, iter = 1000, keep = 50)
}
\arguments{
\item{dtm}{document-term matrix}

\item{k}{number of topics to detect}

\item{seed}{seed for Gibbs LDA}

\item{burn_in}{parameter for Gibbs LDA}

\item{iter}{parameter for Gibbs LDA}

\item{keep}{parameter for Gibbs LDA}
}
\value{
list with elements: lda_out (full LDA topicmodel), beta (beta scores),
gamma (gamma scores), and logLik (Log likelihood of topicmodel object)
}
\description{
Calculate LDA topic model
}
