% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_words.R
\name{stem_words}
\alias{stem_words}
\title{Stem un-nested words before LDA analysis}
\usage{
stem_words(x, exceptions = NULL)
}
\arguments{
\item{x}{character vector of words to be stemmed.}

\item{exceptions}{optional character vector of strings not to stem. Note
these strings must match exactly.}
}
\value{
character vector with their porter stemmed version.
}
\description{
Uses the \code{\link[SnowballC:wordStem]{SnowballC::wordStem()}} function to apply Dr Porter's stemming
algorithm.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

words <- c("win", "winning", "winner", NA)

stem_words(words)
\dontshow{\}) # examplesIf}
}
