% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_stemming.R
\name{summarise_stemming}
\alias{summarise_stemming}
\title{Produce summary table of terms being stemmed ordered by frequency}
\usage{
summarise_stemming(
  data,
  clean_text_fn,
  glossary_words,
  stop_words,
  stem_word_exceptions
)
}
\arguments{
\item{data}{data frame containing response text for all questions.}

\item{clean_text_fn}{function used to clean question text.}

\item{glossary_words}{character vector of multi-word terms to be replaced
with single word equivalents (set to NULL to ignore).}

\item{stop_words}{character vector of common words to be removed from
un-nested values (set to NULL to ignore).}

\item{stem_word_exceptions}{character vector of terms not to stem. Note
these strings must match exactly.}
}
\value{
data frame of stemmed terms
}
\description{
Produce summary table of terms being stemmed ordered by frequency
}
