% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_text_cleaner.R
\name{summarise_text_cleaner}
\alias{summarise_text_cleaner}
\title{Produce summary table of characters replaced by text cleaning function}
\usage{
summarise_text_cleaner(data, clean_text_fn)
}
\arguments{
\item{data}{data frame containing response text for all questions.}

\item{clean_text_fn}{function used to clean question text.}
}
\value{
data frame of character replacements
}
\description{
Produce summary table of characters replaced by text cleaning function
}
