% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_summary.R
\name{summarise_words}
\alias{summarise_words}
\title{Title}
\usage{
summarise_words(
  data,
  text_cleaner,
  min_char = 2,
  stop_words = NULL,
  glossary_words = NULL,
  col_id = NULL,
  cols_free_text = NULL
)
}
\arguments{
\item{cols_free_text}{}
}
\description{
Title
}
