% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_ngrams.R
\name{unnest_ngrams}
\alias{unnest_ngrams}
\title{Un-nest text responses to a question into n-grams for analysis}
\usage{
unnest_ngrams(
  data,
  col_free_text,
  n,
  clean_text_fn,
  glossary_words,
  stop_words,
  col_word = "ngram",
  n_min = n
)
}
\arguments{
\item{data}{data frame containing response text and ids to a single question.}

\item{col_free_text}{character string of column containing the
question text to un-nest.}

\item{n}{integer giving number of words in n-gram (min = 2).}

\item{clean_text_fn}{function used to clean question text.}

\item{glossary_words}{character vector of multi-word terms to be replaced
with single word equivalents (set to NULL to ignore).}

\item{stop_words}{character vector of common words to be removed from
un-nested values (set to NULL to ignore).}

\item{col_word}{character string for un-nested column (default = "ngram")}

\item{n_min}{integer giving minimum number of non-stop words in n-gram to be
included (default = \code{n}).}
}
\value{
data frame of un-nested n-grams.
}
\description{
Note that the passed \code{clean_text_fn} should deal with stripping out
all punctuation, numbers, and changing case as required.
}
