% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_words.R
\name{unnest_words}
\alias{unnest_words}
\title{Unnest text responses to a question into words for analysis by LDA}
\usage{
unnest_words(
  data,
  col_question,
  clean_text_fn,
  glossary_words,
  stop_words,
  stem_word_exceptions = NULL,
  col_word = "word",
  min_char = 2,
  use_stemming = TRUE
)
}
\arguments{
\item{data}{data frame containing question text and response id only.}

\item{col_question}{character string of column containing the
question text to unnest.}

\item{clean_text_fn}{function used to clean question text.}

\item{glossary_words}{set of multi-word character strings to be replaced with
single word equivalents.}

\item{stop_words}{set of common words to be removed from unnested values.}

\item{stem_word_exceptions}{set of character strings not to stem. Note
these strings must match exactly.}

\item{col_word}{character string for unnested column (default = "word")}

\item{min_char}{minimum number of characters needed to be included in
analysis (default = 2)}

\item{use_stemming}{logical indicating if stemming should be applied (if set
to FALSE the value of \code{stem_word_exceptions} is ignored)}
}
\value{
data frame of unnested words
}
\description{
Note that the passed \code{clean_text_fn} should deal with stripping out
all punctuation, numbers, and changing case as required.
}
