% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_summary.R
\name{write_ngram_summary}
\alias{write_ngram_summary}
\title{Helper function to save n-gram summary output to folder}
\usage{
write_ngram_summary(
  ngram_list,
  folder,
  plot_rows = 30,
  prefix = basename(folder)
)
}
\arguments{
\item{ngram_list}{named list of lists where names are n-gram types and
values are named lists where names are questions and values n-gram counts
(generated by \code{\link[=summarise_ngrams]{summarise_ngrams()}}).}

\item{folder}{path to the folder in which to save (will create if not
present).}

\item{plot_rows}{integer number of terms to display on plots.}

\item{prefix}{prefix to use when naming files (defaults to
\code{basename(folder)}).}
}
\value{
passed \code{ngram_list}
}
\description{
Helper function to save n-gram summary output to folder
}
