% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_word_summary.R
\name{write_raw_word_summary}
\alias{write_raw_word_summary}
\title{Title}
\usage{
write_raw_word_summary(data, folder, plot_rows = 30, prefix = basename(folder))
}
\arguments{
\item{prefix}{}
}
\description{
Title
}
